Fix extraction of stored files
When extracting a stored file (no compression), packARC first creates a
temporary file that then renames to the target filename. This file was created
on the directory that stores the archive.
When extracting the files to another filesystem, the rename would fail.
We now create the temporary file on the directory that will hold the file being
extracting, thus, much more likely to be renamed with success to the final name.